[TK006] Merge IsEmpty extensions into root IsEmptyExtensions
Create IsEmptyExtensions.cs at root namespace merging Directory and Drive IsEmpty/HasFiles/HasDirectories. Delete old IsDirectoryEmptyExtensions.cs and DrivesIsEmptyExtensions.cs.
