--- DotExtensions/System/IO/Directories/IsDirectoryEmptyExtensions.cs
@@ -1,80 +1,0 @@
-﻿/*
-        MIT License
-       
-       Copyright (c) 2026 Alastair Lundy
-       
-       Permission is hereby granted, free of charge, to any person obtaining a copy
-       of this software and associated documentation files (the "Software"), to deal
-       in the Software without restriction, including without limitation the rights
-       to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-       copies of the Software, and to permit persons to whom the Software is
-       furnished to do so, subject to the following conditions:
-       
-       The above copyright notice and this permission notice shall be included in all
-       copies or substantial portions of the Software.
-       
-       THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-       IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-       FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-       AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-       LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-       OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-       SOFTWARE.
-   */
-
-using DotExtensions.IO.Files;
-
-namespace DotExtensions.IO.Directories;
-
-/// <summary>
-/// Provides extension methods for directory operations.
-/// </summary>
-public static class IsDirectoryEmptyExtensions
-{
-    private static void ThrowIfDirectoryNotFound(DirectoryInfo directory)
-    {
-        ArgumentNullException.ThrowIfNull(directory);
-
-        if (!directory.Exists)
-            throw new DirectoryNotFoundException(
-                Resources.Exceptions_IO_DirectoryNotFound.Replace("{x}", directory.FullName,
-                    StringComparison.OrdinalIgnoreCase));
-    }
-
-    /// <summary>
-    /// Provides extension methods for checking if a directory is empty.
-    /// </summary>
-    extension(DirectoryInfo directory)
-    {
-        ///<summary>
-        /// Checks if a Directory is empty or not.
-        /// </summary>
-        /// <returns>True if the directory is empty; false otherwise.</returns>
-        /// <exception cref="DirectoryNotFoundException">Thrown if the directory does not exist.</exception>
-        public bool IsEmpty
-        {
-            get
-            {
-                ThrowIfDirectoryNotFound(directory);
-
-                return !directory.SafelyEnumerateFiles().Any() &&
-                       !directory.SafelyEnumerateDirectories().Any();
-            }
-        }
-
-        /// <summary>
-        /// Determines if the directory contains any files.
-        /// </summary>
-        /// <value><see langword="true"/> if the directory has at least one file; otherwise, <see langword="false"/>.</value>
-        /// <exception cref="DirectoryNotFoundException">Thrown if the directory does not exist.</exception>
-        public bool HasFiles
-        {
-            get
-            {
-                ThrowIfDirectoryNotFound(directory);
-
-                return directory.SafelyEnumerateFiles().Any();
-            }
-        }
-    }
-}
