+++ DotExtensions/System/IO/IsEmptyExtensions.cs
@@ -1,5 +1,5 @@
 /*
-        MIT License
+       MIT License
 
        Copyright (c) 2026 Alastair Lundy
 
@@ -22,16 +22,65 @@
        SOFTWARE.
    */
 
-using DotExtensions.IO.Files;
+using DotExtensions.IO.Internal;
 
-namespace DotExtensions.IO.Drives;
+namespace DotExtensions.IO;
 
 /// <summary>
-/// Provides extension methods for working with drive-related operations.
+/// Provides extension properties for checking whether a directory or drive is empty.
 /// </summary>
-public static class DrivesIsEmptyExtensions
+public static class IsEmptyExtensions
 {
-    /// <param name="driveInfo">The drive to be examined.</param>
+    private static void ThrowIfDirectoryNotFound(DirectoryInfo directory)
+    {
+        ArgumentNullException.ThrowIfNull(directory);
+
+        if (!directory.Exists)
+            throw new DirectoryNotFoundException(
+                Resources.Exceptions_IO_DirectoryNotFound.Replace("{x}", directory.FullName,
+                    StringComparison.OrdinalIgnoreCase));
+    }
+
+    /// <summary>
+    /// Provides extension properties for checking whether a directory is empty.
+    /// </summary>
+    extension(DirectoryInfo directory)
+    {
+        /// <summary>
+        /// Checks if a Directory is empty or not.
+        /// </summary>
+        /// <returns>True if the directory is empty; false otherwise.</returns>
+        /// <exception cref="DirectoryNotFoundException">Thrown if the directory does not exist.</exception>
+        public bool IsEmpty
+        {
+            get
+            {
+                ThrowIfDirectoryNotFound(directory);
+
+                return !SafeEnumerator.EnumerateFiles(directory, "*", SearchOption.TopDirectoryOnly, false).Any() &&
+                       !SafeEnumerator.EnumerateDirectories(directory, "*", SearchOption.TopDirectoryOnly, true).Any();
+            }
+        }
+
+        /// <summary>
+        /// Determines if the directory contains any files.
+        /// </summary>
+        /// <value><see langword="true"/> if the directory has at least one file; otherwise, <see langword="false"/>.</value>
+        /// <exception cref="DirectoryNotFoundException">Thrown if the directory does not exist.</exception>
+        public bool HasFiles
+        {
+            get
+            {
+                ThrowIfDirectoryNotFound(directory);
+
+                return SafeEnumerator.EnumerateFiles(directory, "*", SearchOption.TopDirectoryOnly, false).Any();
+            }
+        }
+    }
+
+    /// <summary>
+    /// Provides extension properties for checking whether a drive is empty.
+    /// </summary>
     extension(DriveInfo driveInfo)
     {
         /// <summary>
@@ -42,7 +91,7 @@ public static class DrivesIsEmptyExtensions
             get
             {
                 ArgumentNullException.ThrowIfNull(driveInfo);
-                
+
                 if (!driveInfo.IsReady)
                     return false;
 
@@ -60,8 +109,8 @@ public static class DrivesIsEmptyExtensions
             get
             {
                 ArgumentNullException.ThrowIfNull(driveInfo);
-                
-                return driveInfo.RootDirectory.SafelyEnumerateFiles("*", SearchOption.AllDirectories).Any();
+
+                return SafeEnumerator.EnumerateFiles(driveInfo.RootDirectory, "*", SearchOption.AllDirectories, false).Any();
             }
         }
 
@@ -74,8 +123,8 @@ public static class DrivesIsEmptyExtensions
             get
             {
                 ArgumentNullException.ThrowIfNull(driveInfo);
-                
-                return driveInfo.RootDirectory.SafelyEnumerateDirectories("*", SearchOption.AllDirectories).Any();
+
+                return SafeEnumerator.EnumerateDirectories(driveInfo.RootDirectory, "*", SearchOption.AllDirectories, true).Any();
             }
         }
     }
